Brainfuck

Description

Syntax

data Brainfuck byte addr

Brainfuck language intermediate representation

Constructors

Add byte addr
Add the given value to the memory cell at the given offset from the pointer

Set byte addr
Set the memory cell at the given offset from the pointer to the given value

Mul byte addr addr
Multiply the current offset memory cell and add the result to the combined offset cell

Shift addr
Shift the pointer by some offset

Input addr
Input a byte into the cell at the given offset from the pointer

Output addr
Output the byte at the given offset from the pointer

Loop [Brainfuck byte addr]
Execute the subprogram until the current memory cell is 0

Nop
Do nothing

Bundled Patterns

pattern Dec :: (Eq byte, Eq addr, Num byte, Num addr) => Brainfuck byte addr
Decrement the memory cell at the pointer

pattern In :: (Eq addr, Num addr) => Brainfuck byte addr
Input a char and store it in the memory cell at the pointer

pattern Inc :: (Eq byte, Eq addr, Num byte, Num addr) => Brainfuck byte addr
Increment the memory cell at the pointer

pattern Left_ :: (Eq addr, Num addr) => Brainfuck byte addr
Move the pointer to the left

pattern Right_ :: (Eq addr, Num addr) => Brainfuck byte addr
Move the pointer to the right

pattern Out :: (Eq addr, Num addr) => Brainfuck byte addr
Output the char represented by the memory cell at the pointer

data BrainfuckF (byte :: Type) (addr :: Type) r

Base functor for Brainfuck, for use with recursion schemes

Constructors

AddF byte addr

SetF byte addr

MulF byte addr addr

ShiftF addr

InputF addr

OutputF addr

LoopF [r]

NopF

Bundled Patterns

pattern IncF :: (Eq byte, Eq addr, Num byte, Num addr) => BrainfuckF byte addr a

pattern DecF :: (Eq byte, Eq addr, Num byte, Num addr) => BrainfuckF byte addr a

pattern LeftF_ :: (Eq addr, Num addr) => BrainfuckF byte addr a

pattern RightF_ :: (Eq addr, Num addr) => BrainfuckF byte addr a

pattern InF :: (Eq addr, Num addr) => BrainfuckF byte addr a

pattern OutF :: (Eq addr, Num addr) => BrainfuckF byte addr a

Parsing

parse

Arguments

:: (Num byte, Eq byte, Num addr, Eq addr)

=> FilePath
Source filepath (for error messages)

-> ByteString
Source

-> Either (Diagnostic String) [Brainfuck byte addr]

Parse a brainfuck program from a bytestring

Checks

checks

Arguments

:: forall byte addr. Storable addr

=> RuntimeSettings
Runtime settings

-> [Brainfuck byte addr]
Brainfuck program

-> Maybe (Diagnostic String)

Checks and warnings

Optimization

optimize

Arguments

:: (Num byte, Ord byte, Num addr, Ord addr)

=> Optimization
Optimization options

-> Word
Number of passes

-> [Brainfuck byte addr]
Brainfuck program

-> [Brainfuck byte addr]

Run various optimization passes in sequence

data Optimization

Optimization options

Constructors

Optimization

Fields

contraction :: Bool

deloopification :: Bool

offsets :: Bool

contract :: (Num byte, Ord byte, Num addr, Ord addr) => [Brainfuck byte addr] -> [Brainfuck byte addr]

Contract chains of operations like Add and ShiftR into single combined operations

deloopify :: (Num byte, Eq byte, Num addr, Eq addr) => [Brainfuck byte addr] -> [Brainfuck byte addr]

Replace common loop idioms with single instructions

Configuration

data RuntimeSettings

Settings for brainfuck runtime behavior

Constructors

RuntimeSettings

Fields

memory :: Word64

Number of memory cells available

initialPointer :: Word64

Initial location of the pointer

eofBehavior :: EofBehavior

Behavior when the input stream has reached EOF

data EofBehavior

Action to take when the input stream reaches EOF

Constructors

Zero
Return 0

Unchanged
Leave the cell unchanged

NegativeOne
Return -1

Compilation

compile

Arguments

:: (Integral byte, Storable byte, Integral addr, Storable addr)

=> RuntimeSettings
Runtime settings

-> Word
LLVM optimization level

-> [Brainfuck byte addr]
Brainfuck program

-> IO ByteString

Compile a brainfuck program to object code

jitLLVM

Arguments

:: Word
LLVM optimization level

-> Module
LLVM AST

-> IO ()

Execute an LLVM module with JIT compilation

compileLLVM

Arguments

:: Word
LLVM optimization level

-> Module
LLVM AST

-> IO ByteString

Compile an LLVM module to object code

optimizeLLVM :: Word -> Module -> IO Bool

Optimize an LLVM module

codegen

Arguments

:: forall byte addr. (Integral byte, Storable byte, Integral addr, Storable addr)

=> RuntimeSettings
Runtime settings

-> [Brainfuck byte addr]
Brainfuck program

-> Module

Generate an LLVM module from a brainfuck program

Interpreting

interpretIO

Arguments

:: (Num byte, Eq byte, Storable byte, Unbox byte, Integral addr)

=> Handle
Input handle

-> Handle
Output handle

-> RuntimeSettings
Runtime settings

-> [Brainfuck byte addr]
Brainfuck program

-> IO (IOVector byte, addr)

Interpret in IO, reading from and writing to stdin and stdout

interpret

Arguments

:: (PrimMonad m, Num byte, Eq byte, Unbox byte, Integral addr)

=> (byte -> m byte)
Input function

-> (byte -> m ())
Output function

-> Int
Memory size in bytes

-> addr
Initial pointer location

-> [Brainfuck byte addr]
Brainfuck program

-> m (MVector (PrimState m) byte, addr)

Interpret a brainfuck program

execute

Arguments

:: (PrimMonad m, Num byte, Eq byte, Unbox byte, Integral addr)

=> (byte -> m byte)
Input function

-> (byte -> m ())
Output function

-> MVector (PrimState m) byte
Memory cells

-> MutVar (PrimState m) addr
Pointer

-> Brainfuck byte addr
Brainfuck instruction

-> m ()

Execute a single brainfuck instruction

Pretty Printing

showIR :: (Pretty byte, Eq byte, Num byte, Pretty addr, Eq addr, Num addr) => [Brainfuck byte addr] -> Doc ann

Pretty print brainfuck IR

showLLVM

Arguments

:: Word
LLVM optimization level

-> Module
LLVM AST

-> IO ByteString

Pretty print an LLVM module

showASM

Arguments

:: Word
LLVM optimization level

-> Module
LLVM AST

-> IO ByteString

Pretty print an LLVM module