Brainfuck.Checks

Description

Documentation

memorySize :: forall addr. Storable addr => Word64 -> Maybe (Report String)

Checks if the memory given is larger than the max value of the pointer type or zero

initialPointerPosition :: forall addr. Storable addr => Word64 -> Maybe (Report String)

Checks if the initial pointer is in bounds

heatsTheBox :: [Brainfuck byte addr] -> Maybe (Report String)

Checks if any I/O is performed during the program