Brainfuck.Interpreter

Description

Documentation

interpret

Arguments

:: (PrimMonad m, Num byte, Eq byte, Unbox byte, Integral addr)

=> (byte -> m byte)
Input function

-> (byte -> m ())
Output function

-> Int
Memory size in bytes

-> addr
Initial pointer location

-> [Brainfuck byte addr]
Brainfuck program

-> m (MVector (PrimState m) byte, addr)

Interpret a brainfuck program

execute

Arguments

:: (PrimMonad m, Num byte, Eq byte, Unbox byte, Integral addr)

=> (byte -> m byte)
Input function

-> (byte -> m ())
Output function

-> MVector (PrimState m) byte
Memory cells

-> MutVar (PrimState m) addr
Pointer

-> Brainfuck byte addr
Brainfuck instruction

-> m ()

Execute a single brainfuck instruction